Use reset key approach
